refactor: adjust some attribute types

+++ domainlayer/src/main/java/com/bricklytics/pokesphere/domainlayer/features/pokemon/model/PokemonModel.kt
@@ -12,8 +12,8 @@ data class PokemonModel(
     val isShinny: Boolean = false,
     val officialArtworkModel: PokemonOfficialArtworkModel = PokemonOfficialArtworkModel(),
     val abilities: List<PokemonAbilityModel> = emptyList(),
-    val height: Int = 0,
-    val weight: Int = 0,
+    val height: Double = 0.0,
+    val weight: Double = 0.0,
     val types: List<PokemonTypeModel> = emptyList(),
     val stats: List<PokemonStatModel> = emptyList(),
     val moves: List<PokemonMoveModel> = emptyList()
